Fjerner Api-kallet til Person. Flytter Person-type inn til behandlinger-types, siden det kun er der det blir brukt.
