+++ src/services/modules/behandlinger/behandling/types.ts
@@ -1,6 +1,6 @@
 import { KTObject } from "@navikt/melosys-kodeverk";
 
-import { RegisterAdresse, MidlertidigAdresse, UstrukturertAdresse, Periode, Person, Organisasjon } from "../../types";
+import { RegisterAdresse, MidlertidigAdresse, UstrukturertAdresse, Periode, Organisasjon } from "../../types";
 
 export interface Oppsummering {
   saksnummer: string;
@@ -83,6 +83,27 @@ export interface Medlemsperiode {
   trygdedekning: KTObject;
   kildedokumenttype: KTObject;
   kilde: KTObject;
+}
+
+interface Familiemedlem {
+  fnr: string;
+  sammensattNavn: string;
+  relasjonstype: KTObject;
+  alder: number | null;
+  borMedBruker: boolean;
+  sivilstand: KTObject | null;
+  fnrAnnenForelder: string | null;
+}
+
+interface Person {
+  personStatus: KTObject;
+  sammensattNavn: string;
+  kjoenn: KTObject;
+  fnr: string;
+  foedselsdato: string;
+  sivilstand: KTObject;
+  statsborgerskap: KTObject;
+  familiemedlemmer: Familiemedlem[];
 }
 
 export interface Personhistorikk {
