+++ src/services/modules/types/index.ts
@@ -1,7 +1,6 @@
 import { RegisterAdresse, UstrukturertAdresse, MidlertidigAdresse, StrukturertAdresse } from "./adresse";
 import Fagsak from "./fagsak";
 import Periode from "./periode";
-import Person, { Familiemedlem } from "./person";
 import Organisasjon from "./organisasjon";
 import { Avgiftsberegning, Avgiftsgrunnlag, AvgiftsgrunnlagInfo } from "./trygdeavgift";
 
@@ -15,7 +14,5 @@ export type {
   MidlertidigAdresse,
   StrukturertAdresse,
   Periode,
-  Person,
-  Familiemedlem,
   Organisasjon,
 };
